9PSB VAS API

+++ src/main/java/io/paymentgateway/paymentmodule/NinePsbVasApi/utils/VasData.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class VasData {
 
     private String accessToken;
-    private double expiresIn;
+    private int expiresIn;
 
 }
